Create rust.yml
Just trying out a GitHub workflow that it recommended. It will build and test the project to see if there are any issues whenever I push to master or open a pull request into master.

This could be useful if I make a commit and push without first checking these things myself.